requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
psycopg[binary]==3.2.9
web3==7.13.0
py-solc-x==2.0.4
pytest==8.4.1
httpx==0.28.1
